apps/focus_stacking/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "focus_stacking"
version = "0.1"
description = "Focus stacking application"
requires-python = ">=3.10"
dependencies = ["Pillow>=10.0.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]
